+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and and th
 
 - Disguise the password in the preview of the configuration
 - Used loading dialog for the contexts
+- Wording improved: changed cache to recently loaded
 
 ## 0.0.4